Moved Snake files to correct directory and merged into menu
