+++ CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.24)  # CMake version check
 set(VCPKG_TARGET_ARCHITECTURE x64)
 
 project(ClassicGames)               # Create project 
-set(CMAKE_CXX_STANDARD 20)            # Enable c++14 standard
+set(CMAKE_CXX_STANDARD 20)            # Enable c++20 standard
 
 # Add source files
 set(SOURCE_FILES WordleCode/WordleGameManager.cpp
@@ -19,6 +19,12 @@ connect4/connect4GameManager.cpp
 TicTacToeCode/ttt.cpp
 TicTacToeCode/ttt_functions.cpp
 common/Location.cpp
+SnakeCode/Apple.cpp
+SnakeCode/Snake.cpp
+SnakeCode/SnakeGame.cpp
+SnakeCode/SnakeGameManager.cpp
+SnakeCode/SnakeScreenManager.cpp
+SnakeCode/Timer.cpp
 )
 
 # add .h files
@@ -30,7 +36,13 @@ HangmanCode/Hangman/ScreenManager.h
 TicTacToeCode/TicTacToe/TicTacToe/ttt_functions.h
 WordleCode/Wordle/Wordle/GameManager.h
 WordleCode/Wordle/Wordle/ScreenManager.h
-path.h)
+path.h
+SnakeCode/Apple.h
+SnakeCode/Snake.h
+SnakeCode/SnakeGame.h
+SnakeCode/SnakeGameManager.h
+SnakeCode/SnakeScreenManager.h
+SnakeCode/Timer.h)
 
 # Add executable target with source files listed in SOURCE_FILES variable
 add_executable(ClassicGames ${SOURCE_FILES})
