--- SnakeCode/Snake/Snake.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33403.182
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Snake", "Snake\Snake.vcxproj", "{672BD15B-EE05-4B71-8AF9-3FB34BBF225F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{672BD15B-EE05-4B71-8AF9-3FB34BBF225F}.Debug|x64.ActiveCfg = Debug|x64
-		{672BD15B-EE05-4B71-8AF9-3FB34BBF225F}.Debug|x64.Build.0 = Debug|x64
-		{672BD15B-EE05-4B71-8AF9-3FB34BBF225F}.Debug|x86.ActiveCfg = Debug|Win32
-		{672BD15B-EE05-4B71-8AF9-3FB34BBF225F}.Debug|x86.Build.0 = Debug|Win32
-		{672BD15B-EE05-4B71-8AF9-3FB34BBF225F}.Release|x64.ActiveCfg = Release|x64
-		{672BD15B-EE05-4B71-8AF9-3FB34BBF225F}.Release|x64.Build.0 = Release|x64
-		{672BD15B-EE05-4B71-8AF9-3FB34BBF225F}.Release|x86.ActiveCfg = Release|Win32
-		{672BD15B-EE05-4B71-8AF9-3FB34BBF225F}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {66926D59-C87C-474A-8B41-2C7F54B6310C}
-	EndGlobalSection
-EndGlobal
