--- SnakeCode/Snake/Snake/Location.cpp
@@ -1,6 +1,0 @@
-#include "Location.h"
-Location::Location()
-{
-    this->x = 0;
-    this->y = 0;
-}
