+++ SnakeCode/SnakeGame.h
@@ -1,0 +1,5 @@
+#pragma once
+class SnakeGame {
+public:
+	SnakeGame();
+};
