+++ SnakeCode/SnakeGameManager.cpp
