+++ SnakeCode/SnakeGameManager.h
