--- SnakeGame.h
@@ -1,14 +1,0 @@
-#pragma once
-
-class SnakeGame
-{
-public:
-	SnakeGame();
-	~SnakeGame();
-
-	void GameLoop();
-
-private:
-
-};
-
