--- SnakeLocation.cpp
@@ -1,7 +1,0 @@
-#include "SnakeLocation.h"
-
-SnakeLocation::SnakeLocation()
-{
-    this->x = 0;
-    this->y = 0;
-}
