+++ classic-games.vcxproj.filters
@@ -28,11 +28,11 @@
     <Filter Include="Source Files\common">
       <UniqueIdentifier>{59453502-430c-481d-ab19-7e441359ad71}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Source Files\Snake">
-      <UniqueIdentifier>{60680a58-115a-4bc8-ab3b-369fdac6ceb8}</UniqueIdentifier>
-    </Filter>
     <Filter Include="Source Files\Connect4">
       <UniqueIdentifier>{636e74ac-87e5-439a-8dad-11605fb3720a}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Source Files\Snake">
+      <UniqueIdentifier>{8c3330e9-e73d-432e-bb6c-27f167138d4a}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
@@ -78,25 +78,22 @@
     <ClCompile Include="common\Location.cpp">
       <Filter>Source Files\common</Filter>
     </ClCompile>
-    <ClCompile Include="SnakeGameManager.cpp">
+    <ClCompile Include="SnakeCode\Apple.cpp">
       <Filter>Source Files\Snake</Filter>
     </ClCompile>
-    <ClCompile Include="SnakeScreenManager.cpp">
+    <ClCompile Include="SnakeCode\SnakeGame.cpp">
       <Filter>Source Files\Snake</Filter>
     </ClCompile>
-    <ClCompile Include="Snake.cpp">
+    <ClCompile Include="SnakeCode\Snake.cpp">
       <Filter>Source Files\Snake</Filter>
     </ClCompile>
-    <ClCompile Include="Apple.cpp">
+    <ClCompile Include="SnakeCode\SnakeGameManager.cpp">
       <Filter>Source Files\Snake</Filter>
     </ClCompile>
-    <ClCompile Include="SnakeLocation.cpp">
+    <ClCompile Include="SnakeCode\SnakeScreenManager.cpp">
       <Filter>Source Files\Snake</Filter>
     </ClCompile>
-    <ClCompile Include="Timer.cpp">
-      <Filter>Source Files\Snake</Filter>
-    </ClCompile>
-    <ClCompile Include="SnakeGame.cpp">
+    <ClCompile Include="SnakeCode\Timer.cpp">
       <Filter>Source Files\Snake</Filter>
     </ClCompile>
   </ItemGroup>
@@ -144,7 +141,13 @@
       <Filter>Source Files</Filter>
     </ClInclude>
     <ClInclude Include="common\Location.h">
-      <Filter>Header Files</Filter>
+      <Filter>Source Files\common</Filter>
+    </ClInclude>
+    <ClInclude Include="SnakeCode\Apple.h">
+      <Filter>Source Files\Snake</Filter>
+    </ClInclude>
+    <ClInclude Include="SnakeCode\Snake.h">
+      <Filter>Source Files\Snake</Filter>
     </ClInclude>
     <ClInclude Include="SnakeCode\SnakeGameManager.h">
       <Filter>Source Files\Snake</Filter>
@@ -152,19 +155,7 @@
     <ClInclude Include="SnakeCode\SnakeScreenManager.h">
       <Filter>Source Files\Snake</Filter>
     </ClInclude>
-    <ClInclude Include="SnakeCode\Snake.h">
-      <Filter>Source Files\Snake</Filter>
-    </ClInclude>
-    <ClInclude Include="SnakeCode\Apple.h">
-      <Filter>Source Files\Snake</Filter>
-    </ClInclude>
-    <ClInclude Include="SnakeCode\SnakeLocation.h">
-      <Filter>Source Files\Snake</Filter>
-    </ClInclude>
     <ClInclude Include="SnakeCode\Timer.h">
-      <Filter>Source Files\Snake</Filter>
-    </ClInclude>
-    <ClInclude Include="SnakeCode\SnakeGame.h">
       <Filter>Source Files\Snake</Filter>
     </ClInclude>
   </ItemGroup>
